Refactor shared login data package
