+++ WordPress/src/test/java/org/wordpress/android/sharedlogin/JetpackPublicDataTest.kt
@@ -2,6 +2,7 @@ package org.wordpress.android.sharedlogin
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import org.wordpress.android.sharedlogin.data.JetpackPublicData
 
 class JetpackPublicDataTest {
     private val classToTest = JetpackPublicData()
